requirements.txt
django
mysqlclient
pillow
pandas
openpyxl
django-object-actions
django-wkhtmltopdf
bangla
django-storages
python-dotenv
boto3
djangorestframework
django-cors-headers
psycopg2